move ensure_metadata_has_dp_cp_group to torch/opt/plugins/megatron.py
